library

+++ app/src/main/java/com/example/library/activity/fragment/AccountFragment.kt
@@ -3,7 +3,6 @@ package com.example.library.activity.fragment
 import android.content.Intent
 import android.net.Uri
 import android.os.Bundle
-import android.util.Log
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
@@ -23,7 +22,7 @@ import retrofit2.Call
 import retrofit2.Callback
 import retrofit2.Response
 
-class AccountFragment() : Fragment() {
+class AccountFragment : Fragment() {
     private lateinit var binding: FragmentAccountBinding
     private lateinit var db : AuthDBHelper
     private var userApi = RetrofitService.getInstance().create(UserApi::class.java)
@@ -31,7 +30,7 @@ class AccountFragment() : Fragment() {
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
-    ): View? {
+    ): View {
         // Inflate the layout for this fragment
         binding = FragmentAccountBinding.inflate(inflater, container, false)
         return binding.root
@@ -39,12 +38,13 @@ class AccountFragment() : Fragment() {
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-        db = AuthDBHelper(this.requireContext())
+        db = AuthDBHelper(requireContext())
         val token = db.getToken()
+
         initProfile(token)
 
         binding.btnAccount.setOnClickListener {
-
+            Toast.makeText(requireContext(),"Đang phát triển thêm", Toast.LENGTH_SHORT).show()
         }
 
         binding.btnContact.setOnClickListener {
@@ -54,32 +54,37 @@ class AccountFragment() : Fragment() {
         }
 
         binding.btnWebsite.setOnClickListener {
-            val intent = Intent(Intent.ACTION_DIAL)
+            val intent = Intent(Intent.ACTION_VIEW)
             intent.data = Uri.parse(resources.getString(R.string.website))
             startActivity(intent)
         }
 
         binding.btnLogout.setOnClickListener {
-            val data = authApi.logout("Bearer " + token.accessToken, "Bearer " + token.refreshToken)
-            data.enqueue(object : Callback<Token>{
-                override fun onResponse(call: Call<Token>, response: Response<Token>) {
-                    val result = response.body()
-                    if (result != null){
-                            val rs = db.deleteToken(token.refreshToken!!)
-                            if (rs >= 1){
-                                val intent = Intent(this@AccountFragment.activity, LoginActivity::class.java)
-                                startActivity(intent)
-                                this@AccountFragment.activity?.finish()
-                            }
-                    }
-                }
-
-                override fun onFailure(call: Call<Token>, t: Throwable) {
-                }
-            })
+            logOut(token)
         }
     }
 
+    private fun logOut(token: Token){
+        val data = authApi.logout("Bearer " + token.accessToken, "Bearer " + token.refreshToken)
+        data.enqueue(object : Callback<Token>{
+            override fun onResponse(call: Call<Token>, response: Response<Token>) {
+                if (response.isSuccessful){
+                    val result = response.body()
+                    if (result != null){
+                        val rs = db.deleteToken(token.refreshToken!!)
+                        if (rs >= 1){
+                            val intent = Intent(this@AccountFragment.activity, LoginActivity::class.java)
+                            startActivity(intent)
+                            this@AccountFragment.activity?.finish()
+                        }
+                    }
+                }
+            }
+
+            override fun onFailure(call: Call<Token>, t: Throwable) {
+            }
+        })
+    }
     private fun initProfile(token: Token) {
         val data = userApi.getMyProfile("Bearer " + token.accessToken)
         data.enqueue(object : Callback<Users>{
